Moved the UpperMenu to a seperate component
